Add Sabre models in TravelSample
This commit adds a number of new files under the Sabre namespace in the Infra/Models directory. These files are related to flight segments, origin and destination options, air itinerary pricing info, among others, indicating an increase in the level of detail available in the flight information in the TravelSample application.