Refactor authentication method and expand Amadeus service API
The AmadeusTokenAuthenticationHandler has been updated to improve how the form data for authentication is set. The grant type, client id, and client secret are now added as separate key-value pairs. Addition of two new methods, SearchHotelByCity and GetHotelsData, in IAmadeusService expands the communication with the Amadeus API. The OAuth token method signature was modified in IAmadeusTokenService to include the correct content type header and switch to the Post HTTP verb.